# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Stream Analysis Tools API
  version: 0.3.0
  description: HTTP contract for the stream analysis main API.
servers:
  - url: http://localhost:8080
    description: Local development
tags:
  - name: Operations
    description: Operational endpoints used to observe service health.
  - name: Streams
    description: YouTube stream metadata preview, registration, and retrieval.
  - name: Collections
    description: Background chat-replay collection orchestration and status.
  - name: Chat
    description: Persisted chat messages in stable stream-relative order.
paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Check API health
      description: Reports whether the main API process is ready to accept requests.
      tags:
        - Operations
      responses:
        "200":
          description: The API is healthy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        default:
          $ref: "#/components/responses/Problem"
  /v1/streams/preview:
    post:
      operationId: previewStream
      summary: Preview a YouTube stream
      description: Validates a YouTube URL and fetches current metadata without persisting it.
      tags:
        - Streams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamURLInput"
      responses:
        "200":
          description: Current metadata for the stream.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamPreview"
        "400":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/streams:
    post:
      operationId: createStream
      summary: Register a YouTube stream
      description: Revalidates current YouTube metadata and persists a new stream.
      tags:
        - Streams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamURLInput"
      responses:
        "201":
          description: The stream was registered.
          headers:
            Location:
              description: Relative URL of the registered stream.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    get:
      operationId: listStreams
      summary: List registered streams
      description: Returns registered streams in reverse creation order with offset pagination.
      tags:
        - Streams
      parameters:
        - name: limit
          in: query
          description: Maximum number of streams to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of streams to skip.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A page of registered streams.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamList"
        "400":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/streams/{streamId}:
    get:
      operationId: getStream
      summary: Get a registered stream
      description: Returns one registered stream by its internal identifier.
      tags:
        - Streams
      parameters:
        - name: streamId
          in: path
          required: true
          description: Internal stream identifier.
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
      responses:
        "200":
          description: The registered stream.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/streams/{streamId}/collections:
    post:
      operationId: startCollection
      summary: Start chat-replay collection
      description: >-
        Creates a queued collection job, or returns the already-active job for the stream.
        This operation accepts no request body or collector-specific options.
      tags:
        - Collections
      parameters:
        - $ref: "#/components/parameters/StreamId"
      responses:
        "202":
          description: The new or already-active collection job.
          headers:
            Location:
              description: Relative URL for polling the latest stream collection.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionJob"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/streams/{streamId}/collections/latest:
    get:
      operationId: getLatestCollection
      summary: Get latest collection status
      description: Returns the most recently requested collection job for the stream.
      tags:
        - Collections
      parameters:
        - $ref: "#/components/parameters/StreamId"
      responses:
        "200":
          description: Latest safe collection status and progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionJob"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/collection-jobs/{jobId}/retry:
    post:
      operationId: retryCollection
      summary: Retry a failed collection
      description: >-
        Creates a new queued job only when the failed job has a stable retryable error.
        This operation accepts no request body or collector-specific options.
      tags:
        - Collections
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "202":
          description: The new collection job.
          headers:
            Location:
              description: Relative URL for polling the stream's latest collection.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionJob"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/streams/{streamId}/chat-messages:
    get:
      operationId: listChatMessages
      summary: List persisted chat messages
      description: >-
        Returns messages ordered by offset milliseconds and message ID. The opaque cursor
        contains only the last returned offset and ID.
      tags:
        - Chat
      parameters:
        - $ref: "#/components/parameters/StreamId"
        - name: limit
          in: query
          description: Maximum number of messages to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          description: Opaque cursor returned by the previous page.
          schema:
            type: string
            minLength: 1
            maxLength: 256
      responses:
        "200":
          description: A stable page of chat messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessagePage"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
components:
  parameters:
    StreamId:
      name: streamId
      in: path
      required: true
      description: Internal stream identifier.
      schema:
        type: string
        pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
    JobId:
      name: jobId
      in: path
      required: true
      description: Internal collection job identifier.
      schema:
        type: string
        pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
  responses:
    Problem:
      description: The request could not be completed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required:
        - component
        - status
      properties:
        component:
          type: string
          minLength: 1
          description: Component reporting its health.
          examples:
            - main-api
        status:
          type: string
          enum:
            - ok
          description: Current health status.
    StreamURLInput:
      type: object
      additionalProperties: false
      required:
        - url
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 2048
          description: A supported YouTube watch, live, Shorts, embed, or youtu.be URL.
          examples:
            - https://www.youtube.com/watch?v=dQw4w9WgXcQ
    StreamPreview:
      type: object
      additionalProperties: false
      required:
        - youtubeVideoId
        - canonicalUrl
        - title
        - channelId
        - channelTitle
        - lifecycleStatus
        - metadataFetchedAt
      properties:
        youtubeVideoId:
          type: string
          minLength: 11
          maxLength: 11
        canonicalUrl:
          type: string
          format: uri
        title:
          type: string
          minLength: 1
        channelId:
          type: string
          minLength: 1
        channelTitle:
          type: string
          minLength: 1
        thumbnailUrl:
          type: string
          format: uri
        scheduledStartAt:
          type: string
          format: date-time
        actualStartAt:
          type: string
          format: date-time
        actualEndAt:
          type: string
          format: date-time
        durationMs:
          type: integer
          format: int64
          minimum: 0
        lifecycleStatus:
          $ref: "#/components/schemas/StreamLifecycleStatus"
        metadataFetchedAt:
          type: string
          format: date-time
    Stream:
      allOf:
        - $ref: "#/components/schemas/StreamPreview"
        - type: object
          additionalProperties: false
          required:
            - id
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    StreamList:
      type: object
      additionalProperties: false
      required:
        - items
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Stream"
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
    StreamLifecycleStatus:
      type: string
      enum:
        - unknown
        - scheduled
        - live
        - ended
        - unavailable
    CollectionJob:
      type: object
      additionalProperties: false
      required:
        - id
        - streamId
        - kind
        - status
        - attempt
        - processedCount
        - skippedCount
        - requestedAt
        - updatedAt
      properties:
        id:
          type: string
          pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
        streamId:
          type: string
          pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
        kind:
          type: string
          enum:
            - chat_replay
        status:
          $ref: "#/components/schemas/CollectionJobStatus"
        attempt:
          type: integer
          minimum: 0
        currentStep:
          type: string
          enum:
            - chat_replay
        processedCount:
          type: integer
          format: int64
          minimum: 0
        skippedCount:
          type: integer
          format: int64
          minimum: 0
        requestedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        error:
          $ref: "#/components/schemas/CollectionError"
    CollectionJobStatus:
      type: string
      enum:
        - queued
        - running
        - succeeded
        - no_data
        - failed
    CollectionError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - retryable
      properties:
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]*$"
        message:
          type: string
          minLength: 1
        retryable:
          type: boolean
    ChatMessage:
      type: object
      additionalProperties: false
      required:
        - id
        - authorDisplayName
        - messageText
        - publishedAt
        - offsetMilliseconds
        - messageType
      properties:
        id:
          type: string
          pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
        authorChannelId:
          type: string
          minLength: 1
        authorDisplayName:
          type: string
          minLength: 1
        messageText:
          type: string
          minLength: 1
        publishedAt:
          type: string
          format: date-time
        offsetMilliseconds:
          type: integer
          format: int64
        messageType:
          type: string
          enum:
            - text
    ChatMessagePage:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        nextCursor:
          type: string
          minLength: 1
          maxLength: 256
    ProblemDetails:
      type: object
      additionalProperties: true
      description: RFC 9457 Problem Details object with a stable product-level code.
      required:
        - title
        - status
        - detail
        - code
      properties:
        type:
          type: string
          format: uri-reference
          description: URI reference identifying the problem type; omitted values imply about:blank.
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: HTTP status code generated by the origin server.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying this problem occurrence.
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]*$"
          description: Stable machine-readable product error code.